src/ requirements.txt
pyTelegramBotAPI==4.29.1
python-dotenv==1.0.1
SQLAlchemy==2.0.44
aiohttp==3.13.2
alembic==1.17.2
